if partial is an object, check for parsed and original values
Partials can now be a template string, an array of tokens, or an object.
Using an object for a partial allows specifying the tokens for the template
(.parsed) as well as the original template string (.original). This makes
it possible to use externally cached pre-parsed partial templates that use
higher-order sections.